change return type from `Set<HexColor>` to `HexColor[]`
An object stored in `Set` is compared through by and not by
value. This negates our ability to do membership tests using
`has` as the value being compared will always return false
even when both the objects are identical, but the references
are not the same. So callers of this API now have to test
membership through a linear array scan using `Array.includes`.